refactor(filtered-search): split single and multi select
This PR serves two goals:
1. Decouple logic of different criteria types.
   It helps to prevent unwanted sideeffect of changes and makes the code more readable.
2. Preparing the move of multi select logic from the typeahead to the filtered-search.
   As discussed internally, the typeahead should have never gotten multi-select support
   which anyway only works in the filtered-search.
