+++ projects/element-ng/filtered-search/values/multi-select/si-filtered-search-multi-select.component.scss
@@ -1,0 +1,10 @@
+input {
+  background: transparent;
+  -moz-appearance: textfield; // stylelint-disable-line property-no-vendor-prefix
+
+  &::-webkit-inner-spin-button,
+  &::-webkit-outer-spin-button {
+    -webkit-appearance: none; // stylelint-disable-line property-no-vendor-prefix
+    margin: 0;
+  }
+}
